Remove VLC and QtWebEngine, clear up docs
With the new Rust implementation, VLC suport is dropped. This is mainly
because it doesn't work well with automatic synchronization because it's
not precise enough.

QtWebEngine is also dropped in place of a small web server in the Rust
implementation that asks the user to log in, and automatically obtains
the authentication information. This considerably reduces the size and
complexity of Vidify, since shipping an entire webkit is excessive for
the small usage it had, and anyone has a web browser around anyway.

Vidify is now less modular as well, which means that it's much easier to
distribute and install. It will from now on include support for all APIs
and Players, optionally being able to include new ones manually.